[+FEATURE] Add Google Analytics support and further SEO optimizations
This adds new properties to the Blog model which allow for setting
keywords, a longer description and including a Google Analytics script.

The templates have been optimized a bit to be more SEO friendly. In
particular the blog and post title and description are now used more
intelligently: in a single post view, the post title is used as a <h1>
heading and only the front page uses the blog title as <h1>.

Change-Id: I661a91f1abeafb92a438b689b521e9219d92143d